Function ucd_util::canonical_property_name

pub fn canonical_property_name(
    property_table: PropertyTable, 
    normalized_property_name: &str
) -> Option<&'static str>

Find the canonical property name for the given normalized property name.

If no such property exists, then None is returned.

The normalized property name must have been normalized according to UAX44 LM3, which can be done using symbolic_name_normalize.