Function ucd_util::canonical_property_value

pub fn canonical_property_value(
    property_values: PropertyValues, 
    normalized_property_value: &str
) -> Option<&'static str>

Find the canonical property value for the given normalized property value.

The given property values should correspond to the values for the property under question, which can be found using property_values.

If no such property value exists, then None is returned.

The normalized property value must have been normalized according to UAX44 LM3, which can be done using symbolic_name_normalize.