Function ucd_util::ideograph_name

pub fn ideograph_name(cp: u32) -> Option<String>

Return the character name of the given ideograph codepoint.

This operation is only defined on ideographic codepoints. This includes precisely the following inclusive ranges:

3400..4DB5

4E00..9FD5

20000..2A6D6

2A700..2B734

2B740..2B81D

2B820..2CEA1

17000..187EC

F900..FA6D

FA70..FAD9

2F800..2FA1D

If the given codepoint is not in any of the above ranges, then None is returned.

This implements the algorithm described in Unicode 4.8.