Function ucd_util::property_values

pub fn property_values(
    property_value_table: PropertyValueTable, 
    canonical_property_name: &str
) -> Option<PropertyValues>

Find the set of possible property values for a given property.

The set returned is a mapping expressed as a sorted list of tuples. The first element of each tuple is a normalized property value while the second element of each tuple is the corresponding canonical property value.

If no such property exists, then None is returned.

The given property name must be in its canonical form, which can be found using canonical_property_name.